src/computer/requirements.txt
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
fonttools==4.39.2
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyserial==3.5
pytest==7.2.2
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pyzmq==25.0.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
wcwidth==0.2.6
